README.md includes a section about the sensors, describing the data retreived from it, how they communicate with the server, and how to add one to the server's database.
compensated_sensor_data, t, v and s columns of the sensor_data database table are now optional as the mobile app does not use them, and the former two are unknown. The websocket will also not check for and store the former two values in the database. There is also an explanation of the properties checked for in websocket.js, as well as the format of those properties and why the timestamp is taken from there.

Signed-off-by: n7581769 <st2.smith@hdr.qut.edu.au>