Added a public information page for showing house scores and sensor data to users that are not logged in. Also describes what the readings from the sensor are, and why they are important.
Website is changed to use styling from Fomantic UI (which is included in this repository), to make it appear cleaner for consumption from all users.

Log in form page and login route will give an error if the user is already logged in, and redirect to the home page if it is on the website.

Signed-off-by: n7581769 <st2.smith@hdr.qut.edu.au>